+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Write warning to log if client ID is not provided with OAuth token request.
 - Write warning to log if unsupported OAuth scopes are requested.
 - Return scopes that were requested by the user in `/oauth/token` response.
+- Write warning to log if requested token scopes are not a subset of app scopes.
 
 ### Fixed